feat: Abstract out activity-container
